pyproject.toml
[tool.poetry]
name = "snkrs-news-bot"
version = "0.1.0"
description = ""
authors = ["Shuntaro Shimizu <ut.s.shimizu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.15.0"
autopep8 = "^1.5.4"
selenium = "^3.141.0"
pandas = "^1.1.4"
slackbot = "^1.0.0"
slacker = "^0.14.0"
beautifulsoup4 = "^4.9.3"
urllib3 = "^1.26.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
